fix: thermal: fix negative temperature calculation
negative temperature value was wrong and not all variables (i.e. the
register value) were represented as 'signed' in the calculation.

Reviewed-on: http://vgitil04.il.marvell.com:8080/32385
Tested-by: Star_Automation <star@marvell.com>

+++ drivers/thermal/mvebu_ext_thermal_28nm.c
@@ -62,7 +62,7 @@ s32 mvebu_thermal_ext_sensor_read(struct thermal_unit_config *tsen)
 	if (reg >= THERMAL_SEN_OUTPUT_MSB)
 		reg -= THERMAL_SEN_OUTPUT_COMP;
 
-	return ((tsen->tsen_gain * reg) + tsen->tsen_offset) / tsen->tsen_divisor;
+	return ((tsen->tsen_gain * ((s32)reg)) + tsen->tsen_offset) / tsen->tsen_divisor;
 }
 
 u32 mvebu_thermal_ext_sensor_probe(struct thermal_unit_config *tsen)
